refactor: apply scream architecture ==> https://blog.cleancoder.com/uncle-bob/2011/09/30/Screaming-Architecture.html
