+++ src/application/date-provider.ts
@@ -1,0 +1,3 @@
+export interface DateProvider {
+    getNow(): Date;
+}
