+++ src/application/domain/message.ts
