+++ src/infra/real-date-provider.ts
@@ -1,0 +1,7 @@
+import {DateProvider} from "../application/date-provider";
+
+export class RealDateProvider implements DateProvider {
+    getNow(): Date {
+        return new Date();
+    }
+}
