+++ tsconfig.json
@@ -3,6 +3,8 @@
   "compilerOptions": {
     "target": "es5",
     "module": "commonjs",
+    "noEmit": true,
+    "downlevelIteration": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
